requirements.txt
django
django_rest_framework
mysqlclient
djangorestframework-simplejwt